fix(front) : 빌드시 NextAuth typeError가 나는 문제 수정
